requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
jsonschema==4.17.3
mysqlclient==2.1.1
pyrsistent==0.19.3
pytz==2022.7.1
sqlparse==0.4.3
